écriture sur la BDD via runner

+++ src/main/java/fr/eni/TPfilmotheque/controller/StringToGenreConverter.java
@@ -20,7 +20,7 @@ public class StringToGenreConverter implements Converter<String, Genre>{
 	
 	@Override
 	public Genre convert(String id) {
-		Integer theId = Integer.parseInt(id);
+		Long theId = Long.parseLong(id);
 		
 		Genre test = service.getMapGenres().get(theId);
